pyproject.toml
[project]
name = "legal-automation-os"
version = "1.0.0"
description = "Jurisdiction-aware compliance document generation engine"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "jsonschema>=4.20",
    "rich>=13.0",
    "click>=8.1",
    "python-docx>=1.0",
    "pypandoc>=1.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.1",
]

[project.scripts]
legal-os = "engine.cli:main"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
